==> Library/Mobile Documents/com~apple~CloudDocs/Documents/TBWA/challenge-Amazon-main/data/processing/04_data_processing_book_info.ipynb <==
## Dataset: book_info

import pandas as pd

### Import dataset

df = pd.read_parquet("data/raw/book_info.parquet")

### Data cleaning

# remove colum
df = df.drop(columns=["description"])

# joining a list of strings into a single string
df["authors"] = df["authors"].apply(lambda x: ", ".join(x))

df["categories"] = df["categories"].apply(lambda x: ", ".join(x))

# remove rows where contain "not available"
mask = ~df["authors"].str.contains("not available", na=False) & ~df[
    "categories"
].str.contains("not available", na=False)

df_book = df.loc[mask].copy()

### Exploratory data analysis


df_book.info()  # data type

df_book[df_book["authors"].str.contains("not available", case=False)]

df_book["title"].value_counts().head(5)

# the same title represents different editions/volumes
df_book[df_book["title"].str.contains("case closed", case=False)]

# authored the most number of  books
df_book.groupby("authors")["title"].count().nlargest(10)

# bottom 5 authors
df_book["authors"].value_counts().tail(5)

# top 5 categories
df_book["categories"].value_counts().head(5)

# the categories are inconsistent?
df_book.loc[
    df_book["title"] == "Harry Potter", ["isbn", "title", "categories", "authors"]
]

# categories: England not seem to be correct
df_book.loc[
    df_book["title"].str.startswith("Harry Potter and the Prisoner of Azkaban")
    & (df_book["authors"] == "J. K. Rowling"),
    ["isbn", "title", "categories", "authors"],
]

df_book.loc[
    df_book["categories"] == "England", ["isbn", "title", "categories", "authors"]
]

### Merge dataset: purchases

# select columns
columns = [
    "Survey ResponseID",
    "ASIN/ISBN (Product Code)",
    "Category",
    "Quantity",
    "Purchase Price Per Unit",
    "Purchase Total",
]
df_purchases = pd.read_parquet("data/ready/amazon_purchases.parquet", columns=columns)

# Filter by ISBN with 10 and 13 digits and Category Abis book
df_purchases = df_purchases[
    (df_purchases["Category"] == "Abis book")
    & (df_purchases["ASIN/ISBN (Product Code)"].str.len().isin([10, 13]))
]

df_book_purchases = pd.merge(
    df_book,
    df_purchases,
    left_on="isbn",
    right_on="ASIN/ISBN (Product Code)",
    how="inner",
)
df_book_purchases = df_book_purchases.drop(columns=["ASIN/ISBN (Product Code)"])

df_book_purchases.info()

### Merge dataset: survey

# select columns
columns = [
    "Survey ResponseID",
    "Q-demos-gender",
    "Q-demos-age",
    "Q-demos-education",
    "Q-demos-state",
    "Q-demos-income",
]

df_survey = pd.read_parquet("data/ready/amazon_survey.parquet", columns=columns)

df_book_purchases_survey = pd.merge(
    df_book_purchases,
    df_survey,
    on="Survey ResponseID",
    how="inner",
)

# top 5 categories
df_book_purchases_survey["categories"].value_counts().head(5)

# top 5 title
df_book_purchases_survey["title"].value_counts().head(5)

# best selling authors
df_book_purchases_survey.groupby("authors")["Survey ResponseID"].nunique().nlargest(10)

df_book_purchases_survey["Survey ResponseID"].nunique()

# book purchases of a specific user
df_book_purchases_survey[
    df_book_purchases_survey["Survey ResponseID"] == "R_phoOTxhiHcWo7lL"
]

# top 10 users with the highest books purchase total amounts
top10_survey = (
    df_book_purchases_survey.groupby("Survey ResponseID")["Purchase Total"]
    .sum()
    .nlargest(10)
    .index
)

# filter dataframe with top 10 users
df_top10_book_purchases_survey = df_book_purchases_survey[
    df_book_purchases_survey["Survey ResponseID"].isin(top10_survey)
]

df_top10_book_purchases_survey.head(1)

### Save dataset

selected_columns = [
    "Survey ResponseID",
    "title",
    "authors",
    "categories",
    "Q-demos-gender",
    "Q-demos-age",
    "Q-demos-education",
    "Q-demos-state",
    "Q-demos-income",
]
df_top10_book_purchases_survey = df_top10_book_purchases_survey[
    selected_columns
].reset_index(drop=True)

#### Grouping Survey Data: Combining Demographics and Book Purchases

df_top10_book_purchases_survey.info()

# unique IDs
unique_ids = df_top10_book_purchases_survey["Survey ResponseID"].unique()

# create separate dictionaries for demographic and book purchases
demographic_data = {}
book_data = {survey_id: [] for survey_id in unique_ids}

# iterate over the dataframe rows
for _, row in df_top10_book_purchases_survey.iterrows():
    survey_id = row["Survey ResponseID"]

    # extract demographic information
    demographic_info = {
        "Gender": row["Q-demos-gender"],
        "Age": row["Q-demos-age"],
        "Education level": row["Q-demos-education"],
        "State": row["Q-demos-state"],
        "Household income": row["Q-demos-income"],
    }

    # extract book purchases
    book_info = {
        "authors": row["authors"],
        "categories": row["categories"],
        "title": row["title"],
    }

    # store demographic information (once per ID)
    if survey_id not in demographic_data:
        demographic_data[survey_id] = demographic_info

    # add purchase information to the corresponding group in the dictionary
    book_data[survey_id].append(book_info)

# combine informations into a DataFrame
combined_data = []
for survey_id in unique_ids:
    user_info = demographic_data[survey_id]
    books = book_data[survey_id]
    combined_data.append(
        {
            "Survey ResponseID": survey_id,
            "demographic": user_info,
            "books": books,
        }
    )

# converter do df
df_grouped = pd.DataFrame(combined_data)


# save json file
df_grouped.to_json("data/ready/top10_grouped_book_purchases.json")
